Relocate component to `UI` folder
